mejorando las opciones hhttp sobre usuarios

+++ apps/users/api/api.py
@@ -1,41 +1,51 @@
+from rest_framework import status
 from rest_framework.response import Response
 from rest_framework.views import APIView
 from rest_framework.decorators import api_view
 from apps.users.api.serializers import UserSerializer
 from apps.users.models import User
 
+
 @api_view(['GET','POST'])
 def user_api_view(request):
 
+    #list
     if request.method == 'GET':
         users = User.objects.all()
         users_serializer = UserSerializer(users,many = True)
-        return Response(users_serializer.data)
+        return Response(users_serializer.data,status = status.HTTP_200_OK)
 
+    #create
     elif request.method == 'POST':
         user_serializer = UserSerializer(data = request.data)
         if user_serializer.is_valid():
             user_serializer.save()
-            return Response(user_serializer.data)
-        return Response(user_serializer.errors)
+            return Response({'message':'Usuario creado correctamente.'}, status = status.HTTP_201_CREATED)
+        return Response({'message':'Usuario eliminado correctamente.'}, status = status.HTTP_400_BAD_REQUEST)
 
 @api_view(['GET','PUT','DELETE'])
 def user_detail_api_view(request, pk):
+    user = User.objects.filter(id = pk).first()
 
-    if request.method == 'GET':
-        user = User.objects.filter(id = pk).first()
-        user_serializer = UserSerializer(user)
-        return Response(user_serializer.data)
+    #validation
+    if user:
+        
+        #retrive
+        if request.method == 'GET':
+            user_serializer = UserSerializer(user)
+            return Response(user_serializer.data, status = status.HTTP_200_OK)
 
-    elif request.method == 'PUT':
-        user = User.objects.filter(id = pk).first()
-        user_serializer = UserSerializer(user,data = request.data)
-        if user_serializer.is_valid():
-            user_serializer.save()
-            return Response(user_serializer.data)
-        return Response(user_serializer.errors)
+        #update
+        elif request.method == 'PUT':
+            user_serializer = UserSerializer(user,data = request.data)
+            if user_serializer.is_valid():
+                user_serializer.save()
+                return Response(user_serializer.data, status = status.HTTP_200_OK)
+            return Response(user_serializer.errors, status = status.HTTP_400_BAD_REQUEST)
 
-    elif request.method == 'DELETE':
-        user = User.objects.filter(id = pk).first()
-        user.delete()
-        return Response('Eliminado')
+        #delete
+        elif request.method == 'DELETE':
+            user.delete()
+            return Response({'message':'Usuario eliminado correctamente.'},status = status.HTTP_200_OK)
+    
+    return Response({'message':'Usuario no encontrado.'},status = status.HTTP_404_NOT_FOUND)
